Move code to injectPrFilesSearch
